spark-hdfs-mongo/requirements.txt
pymongo
streamlit
pandas
Kafka
MongoDB
Streamlit
Hadoop
Spark

